+++ R/utils.R
@@ -31,17 +31,11 @@ fix_0haz <- function(haz_est) {
 ##' @export
 ralife_cdf <- function(n, x) {
   if ("event_type" %in% names(x)) {
-    etypes <- unique(x$event_type)
-    out <- vector(mode = "list",
-                  length = length(etypes))
-    for (i in seq_along(etypes)) {
-      x_sub <- x[x$event_type == etypes[i], ]
-      u <- stats::runif(n)
-      indices <- pmin(findInterval(u, x_sub$cdf) + 1, NROW(x_sub))
-      out[[i]] <- data.frame(event_type = etypes[i],
-                             lifetime = x_sub$lifetime[indices])
-    }
-    out <- do.call(rbind, out)
+    lookup <- seq_len(NROW(x))
+    id_samples <- sample(x = lookup, size = n, prob = x$density,
+                         replace = TRUE)
+    out <- data.frame(event_type = x$event_type[id_samples],
+                      lifetime   = x$lifetime[id_samples])
   } else {
     u <- stats::runif(n)
     indices <- pmin(findInterval(u, x$cdf) + 1, NROW(x))