+++ NEWS.md
@@ -1,5 +1,12 @@
 # abslife 0.2.0
 
+- Making `ev_life` compatible with competing risks.
+
+- Reporting density for cases where `hazard = 1` as well.
+
+- Fixed functions for plotting multiple confidence intervals (using logit
+  instead of exp)
+
 - Function for calculating APV (single loan)
 
 # abslife 0.1.5