+++ NEWS.md
@@ -1,5 +1,10 @@
 # abslife 0.2.0
 
+- Minor fix to make base `plot` and `ggauto` methods for competing risk display
+  event types in the same order.
+
+- The warning message for `estimate_hazard` has been updated
+
 - Making `ev_life` compatible with competing risks.
 
 - Reporting density for cases where `hazard = 1` as well.