+++ NEWS.md
@@ -1,5 +1,9 @@
 # abslife 0.2.0
 
+- Function `ralife_cdf` now works with competing risks as well.
+
+- Included `aloans_recov` dataset with recovery curves.
+
 - Minor fix to make base `plot` and `ggauto` methods for competing risk display
   event types in the same order.